+++ App/FrmRecognizerImage.cpp
@@ -54,7 +54,7 @@ CFrmRecognizerImage::~CFrmRecognizerImage()
 
 void CFrmRecognizerImage::on_pbBrower_clicked()
 {
-    QString szFile = RabbitCommon::CDir::GetOpenFileName(this,
+    QString szFile = QFileDialog::getOpenFileName(this,
                                                          tr("Select image"),
                                                          ui->leFile->text(),
            tr("Image files(*.png *.gif *.jpeg *.jpg *.bmp);; All files(*.*)"));