+++ ChangeLog.md
@@ -9,6 +9,7 @@
     See RabbitCommon: 7ed300a206d8a36c30003b6c19ad181682d0ff5f
   + FFmpeg: update API avpicture_* to av_image_*
   + Support Qt6
+  + Replace RabbitCommon::CDir::GetOpenDirectory with QFileDialog::getExistingDirectory etc
 
 - v0.0.4
   + Add opencv face